feat(user deletion): improve delete-account page
- add a navtop title (entry in $lib/navigation)
- improve CSS (center stuff)
- add paragraph at bottom to show support email (useful when e.g.
webhook fails)

+++ packages/app/src/routes/(app)/delete-account/+page.svelte
@@ -1,8 +1,10 @@
 <script lang="ts">
   import { goto } from '$app/navigation';
+  import { env } from '$env/dynamic/public';
   import { graphql } from '$houdini';
   import ButtonSecondary from '$lib/components/ButtonSecondary.svelte';
   import InputText from '$lib/components/InputText.svelte';
+  import LoadingScreen from '$lib/components/LoadingScreen.svelte';
   import MaybeError from '$lib/components/MaybeError.svelte';
   import { LoadingText } from '$lib/loading';
   import { mutateAndToast } from '$lib/mutations';
@@ -27,13 +29,16 @@
 
   let myUid = '';
   let step: 'initial' | 'confirm-uid' = 'initial';
+  let loggingOut = false;
+  let deleting = false;
 </script>
 
 <MaybeError result={$PageDeleteAccount} let:data={{ me }}>
   <div class="contents">
-    <h1>Supprimer ton compte</h1>
-    {#if step === 'initial'}
-      <p>Es-tu sûre, <LoadingText value={me.fullName} /> ?</p>
+    {#if loggingOut}
+      <LoadingScreen title="Déconnexion…" />
+    {:else if step === 'initial'}
+      <p>Es-tu sûre de vouloir supprimer ton compte, <LoadingText value={me.fullName} /> ?</p>
       {#if me.deletionConsequences.length > 0}
         <p>Supprimer ton compte supprimerait...</p>
         <ul>
@@ -46,34 +51,66 @@
       <p>Rentre ton identifiant pour confirmer la suppression</p>
       <InputText label="Identifiant" bind:value={myUid} />
     {/if}
-    <nav>
-      <ButtonSecondary href={route('/settings')}>Non, annuler</ButtonSecondary>
-      <ButtonSecondary
-        disabled={step === 'confirm-uid' && myUid !== me.uid}
-        danger
-        on:click={async () => {
-          if (step === 'initial') {
-            step = 'confirm-uid';
-            return;
-          }
+    {#if !loggingOut}
+      <nav>
+        <ButtonSecondary href={route('/settings')}>Non, annuler</ButtonSecondary>
+        <ButtonSecondary
+          loading={deleting}
+          disabled={step === 'confirm-uid' && myUid !== me.uid}
+          danger
+          on:click={async () => {
+            if (step === 'initial') {
+              step = 'confirm-uid';
+              return;
+            }
 
-          const ok = await mutateAndToast(DeleteAccount, null, {
-            success: 'Compte supprimé 🫡',
-            error: 'Erreur lors de la suppression du compte',
-          });
-          if (ok) {
-            window.localStorage.clear();
-            window.sessionStorage.clear();
-            setTimeout(() => goto(route('/logout')), 2000);
-          }
-        }}>Oui</ButtonSecondary
-      >
-    </nav>
+            deleting = true;
+            const ok = await mutateAndToast(DeleteAccount, null, {
+              success: 'Compte supprimé 🫡',
+              error: 'Erreur lors de la suppression du compte',
+            });
+            deleting = false;
+            if (ok) {
+              window.localStorage.clear();
+              window.sessionStorage.clear();
+              loggingOut = true;
+              setTimeout(
+                () =>
+                  goto(
+                    route('/logout', {
+                      userWasDeleted: true,
+                    }),
+                  ),
+                2000,
+              );
+            }
+          }}>Oui</ButtonSecondary
+        >
+      </nav>
+
+      <footer>
+        <p>Quelque chose ne va pas?</p>
+        <p>
+          Tu peux toujours contacter le support en écrivant à
+          <a href="mailto:{env.PUBLIC_SUPPORT_EMAIL}">{env.PUBLIC_SUPPORT_EMAIL}</a>
+        </p>
+      </footer>
+    {/if}
   </div>
 </MaybeError>
 
 <style>
   .contents {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+    justify-content: center;
     padding: 0 1rem;
+    text-align: center;
+  }
+
+  footer {
+    margin: 4rem 3rem;
+    text-align: center;
   }
 </style>
